refactor: move knex files to a knex module in infra layer
