--- src/infra/db-postgresql/config.ts
@@ -1,6 +1,0 @@
-export = {
-  host: 'localhost',
-  user: 'admin',
-  password: 'admin',
-  database: 'genericSign'
-}
