+++ src/infra/knex/knex-environment.ts
@@ -1,6 +1,6 @@
 import Knex from 'knex'
 const environment = process.env.NODE_ENV || 'production'
 // eslint-disable-next-line @typescript-eslint/no-var-requires
-const config: Knex.Config = require('../../../knexfile')[environment]
+const config: Knex.Config = require('./knexfile')[environment]
 
 export default Knex(config)
